requirements.txt
openai>=1.40.0
xai-sdk>=1.5.0
yfinance>=0.2.40
python-dotenv>=1.0.1
tweepy>=4.0.0
requests>=2.31.0
tenacity>=8.2.3
feedgen>=1.0.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
Pillow>=10.0.0
websockets>=12.0
pytest>=7.0.0
pyyaml>=6.0
jinja2>=3.1.0
boto3>=1.28.0
faster-whisper>=1.0.0
google-api-python-client>=2.120.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0